Created hallucination_detector_agent
this hallucination detector agent is a submission by the team 'The Heckers' for the hackathon Frosthack.

This hallucination detector currently is hardcoded with a sample question and uses llama to generate a responses, the duckduckgo_search_agent to scrape the correct answer and deepseek-r1 to check the validity to give one of the 3 outputs: Correct, Uncertain or Hallucinated. The agent works well for different questions and is well tested. It can be modified to take inputs and give back correct answers.